+++ feature/news/src/main/java/com/teamwable/news/news/contract/NewsInfoSideEffect.kt
@@ -1,8 +1,9 @@
-package com.teamwable.news.news.model
+package com.teamwable.news.news.contract
 
+import com.teamwable.common.base.SideEffect
 import com.teamwable.model.news.NewsInfoModel
 
-sealed interface NewsInfoSideEffect {
+sealed interface NewsInfoSideEffect : SideEffect {
     data class NavigateToDetail(val newsInfoModel: NewsInfoModel) : NewsInfoSideEffect
 
     data class ShowSnackBar(val message: String) : NewsInfoSideEffect
